# Conflicts:
#	app/cart/go.mod
#	app/cart/go.sum
#	app/hertz_gateway/biz/service/checkout.go
#	app/hertz_gateway/biz/service/checkout_waiting.go
#	app/hertz_gateway/biz/service/get_cart.go
#	app/hertz_gateway/biz/utils/resp.go
#	app/hertz_gateway/infra/rpc/client.go
#	app/product/biz/model/category.go
#	idl/product.proto

+++ app/hertz_gateway/biz/service/checkout_waiting.go
@@ -24,7 +24,7 @@ func NewCheckoutWaitingService(Context context.Context, RequestContext *app.Requ
 func (h *CheckoutWaitingService) Run(req *checkout.CheckoutReq) (resp map[string]any, err error) {
 	userId := frontendutils.GetUserIdFromCtx(h.Context)
 	_, err = rpc.CheckoutClient.Checkout(h.Context, &rpccheckout.CheckoutReq{
-		UserId:    uint32(userId),
+		UserId:    userId,
 		Email:     req.Email,
 		Firstname: req.Firstname,
 		Lastname:  req.Lastname,
